+++ tests/cpp/test_sampler.cc
@@ -1,0 +1,18 @@
+#include <gtest/gtest.h>
+#include <dgl/sample_utils.h>
+#include "./common.h"
+
+using namespace dgl;
+using namespace dgl::runtime;
+
+TEST(SampleUtilsTest, TestAliasSampler) {
+
+};
+
+TEST(SampleUtilsTest, TestCDFSampler) {
+
+};
+
+TEST(SampleUtilsTest, TestTreeSampler) {
+
+};
